backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
psycopg[binary]==3.2.7
python-multipart==0.0.20
passlib==1.7.4
pyjwt==2.10.1
pydantic[email]==2.11.3
pydantic-settings==2.8.1
scikit-learn==1.5.2
numpy==2.1.3
pandas==2.2.3
